requirements.txt
streamlit
pandas
pdfplumber
gspread
oauth2client
streamlit-cropper
Pillow
fpdf